Add first React testing library test to get setup
