+++ package.json
@@ -15,6 +15,8 @@
     "@fortawesome/free-regular-svg-icons": "^5.12.1",
     "@fortawesome/free-solid-svg-icons": "^5.8.1",
     "@fortawesome/react-fontawesome": "^0.1.4",
+    "@testing-library/jest-dom": "^5.10.1",
+    "@testing-library/react": "^10.2.1",
     "axios": "^0.19.0",
     "axios-mock-adapter": "^1.15.0",
     "babel-jest": "^24.5.0",
